+++ automotive/src/main/java/de/moleman1024/audiowagon/ServicePriority.kt
@@ -1,0 +1,7 @@
+package de.moleman1024.audiowagon
+
+enum class ServicePriority {
+    BACKGROUND,
+    FOREGROUND_REQUESTED,
+    FOREGROUND
+}